+++ blazeutils/xlrd.py
@@ -1,18 +1,11 @@
-from __future__ import absolute_import
+import warnings
 
-try:
-    from cStringIO import StringIO
-except ImportError:
-    from StringIO import StringIO
+from .spreadsheets import workbook_to_reader as _workbook_to_reader
 
-import xlrd
+warnings.warn('xlrd module deprecated, workbook_to_reader() moved to spreadsheets module',
+              DeprecationWarning)
 
 def workbook_to_reader(xlwt_wb):
-    """
-        convert xlwt Workbook instance to an xlrd instance for reading
-    """
-    fh = StringIO()
-    xlwt_wb.save(fh)
-    # prep for reading
-    fh.seek(0)
-    return xlrd.open_workbook(file_contents=fh.read())
+    _workbook_to_reader(xlwt_wb)
+
+