+++ tests/test_xlrd.py
@@ -1,16 +1,8 @@
-from nose.tools import eq_
-import xlwt
+from blazeutils.testing import emits_deprecation
 
-from blazeutils.xlrd import workbook_to_reader
 
 class TestWorkbookToReader(object):
 
-    def test_it(self):
-
-        write_wb = xlwt.Workbook()
-        ws = write_wb.add_sheet('Foo')
-        ws.write(0,0,'bar')
-
-        wb = workbook_to_reader(write_wb)
-        sh = wb.sheet_by_name('Foo')
-        eq_(sh.cell_value(rowx=0,colx=0), 'bar')
+    @emits_deprecation('xlrd module deprecated, workbook_to_reader\(\) moved to spreadsheets module')
+    def test_deprecation(self):
+        from blazeutils.xlrd import workbook_to_reader